chore: moved back pixel data to document
