+++ pages/_app.tsx
@@ -2,12 +2,8 @@ import { useEffect } from "react"
 
 import "@/styles/globals.scss"
 import { AppPropsWithLayout } from "@/types/Layout.types"
-import { useRouter } from "next/router"
-import * as fbq from '../lib/fb-pixel'
-import Script from "next/script"
-import Head from "next/head"
+
 function MyApp({ Component, pageProps }: AppPropsWithLayout) {
-  const router = useRouter()
 
   useEffect( () => {
     // we need import elements with commonJS
@@ -15,19 +11,6 @@ function MyApp({ Component, pageProps }: AppPropsWithLayout) {
       require("@lottus23/lottus-elements-uane/elements")
     }
   }, [])
-  useEffect(() => {
-    // This pageview only triggers the first time (it's important for Pixel to have real information)
-    // fbq.pageview()
-
-    const handleRouteChange = () => {
-      fbq.pageview()
-    }
-
-    router.events.on('routeChangeComplete', handleRouteChange)
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange)
-    }
-  }, [router.events])
 
 
   // Use the layout defined at the page level, if available
@@ -35,34 +18,6 @@ function MyApp({ Component, pageProps }: AppPropsWithLayout) {
 
   return getLayout(
     <>
-    <Head>
-      <noscript>
-        <img
-          height="1"
-          width="1"
-          style={{ display: 'none' }}
-          src={`https://www.facebook.com/tr?id=487461332128996&ev=PageView&noscript=1`}
-        />
-      </noscript>
-    </Head>
-      <Script
-        id="fb-pixel"
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-            !function(f,b,e,v,n,t,s)
-            {if(f.fbq)return;n=f.fbq=function(){n.callMethod?
-            n.callMethod.apply(n,arguments):n.queue.push(arguments)};
-            if(!f._fbq)f._fbq=n;n.push=n;n.loaded=!0;n.version='2.0';
-            n.queue=[];t=b.createElement(e);t.async=!0;
-            t.src=v;s=b.getElementsByTagName(e)[0];
-            s.parentNode.insertBefore(t,s)}(window, document,'script',
-            'https://connect.facebook.net/en_US/fbevents.js');
-            fbq('init', 487461332128996);
-            fbq('track', 'PageView');
-          `,
-        }}
-      />
       <Component {...pageProps} />
     </>)
 }
